+++ CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to
 - 🚸(backend) improve users similarity search and sort results #391
 - 🌐(backend) add german translation #259
 
+## Fixed
+
+- 🦺(backend) add comma to sub regex #408
+
+
 ## [1.7.0] - 2024-10-24
 
 ## Added